refactoring code between Scene and ABA
