+++ tests/tests/test_integration/test_main.py
@@ -77,9 +77,11 @@ def test_camera():
 
     # render
     scene.render(interactive=False, )
+    scene.close()
 
     # Now render but with a different view
     scene.render(interactive=False, camera='sagittal', zoom=1)
+    scene.close()
 
     # Now render but with specific camera parameters
     bespoke_camera = dict(
@@ -144,6 +146,7 @@ def test_mouselight():
 
     scene.add_brain_regions(['MOs'], alpha=0.15) 
     scene.render(interactive=False, camera='coronal') 
+    scene.close()
 
 
 def test_scene_title():
@@ -166,6 +169,7 @@ def test_streamlines():
     scene.add_streamlines(data, color="darkseagreen", show_injection_site=False)
 
     scene.render(interactive=False, camera='sagittal', zoom=1)
+    scene.close()
 
 def test_tractography():
     from brainrender.Utils.ABA.connectome import ABA
@@ -184,6 +188,7 @@ def test_tractography():
     scene.add_tractography(tract, display_injection_structure=False, color_by="region")
 
     scene.render(interactive=False, )
+    scene.close()
 
 def test_video():
     from brainrender.Utils.videomaker import VideoMaker
